trial: do not merge Changed code to not construct the path from user-controlled data.
Issue-ID: VNFSDK-834

+++ vnfmarket-be/vnf-sdk-marketplace/src/main/java/org/onap/vtp/scenario/VTPScenarioResource.java
@@ -431,8 +431,10 @@ public class VTPScenarioResource extends VTPResource{
         }
 
         try {
-            FileUtils.deleteQuietly(new File(VTP_YAML_STORE, scenarioName));
-            FileUtils.deleteDirectory(scenarioDir);
+            if(FileUtils.directoryContains(new File(VTP_YAML_STORE), new File(scenarioName))) {
+                FileUtils.deleteQuietly(new File(VTP_YAML_STORE, scenarioName));
+            }
+            FileUtils.deleteDirectory(scenarioDir);//test
         } catch (IOException e) {
             LOG.error("Delete scenario yaml {} failed", scenarioName, e);
             throw new VTPException(
